pull recipient & author up into MailNotificationJob
